backend/requirements.txt
Flask==3.1.1
Flask-Cors==5.0.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
bcrypt==4.3.0
PyJWT==2.12.0
Flask-Limiter==3.12.0
gunicorn==23.0.0
